pyproject.toml
[tool.poetry]
name="Trip-Planner"
version="0.1.0"
description="A simple trip planner"
authors = ["Vansh Bhavsar"]



[tool.poetry.dependencies]
python = ">=3.10,<3.13"
crewai = "^0.102.0"
unstructured = '==0.10.25'
tools='0.1.9'
python-dotenv='1.0.0'
langchain-community = "^0.3.18"
langchain-groq = "^0.2.4"
groq = "^0.18.0"

[tool.pyright]
# https://github.com/microsoft/pyright/blob/main/docs/configuration.md
useLibraryCodeForTypes = true
exclude = [".cache"]

[tool.ruff]
# https://beta.ruff.rs/docs/configuration/
select = ['E', 'W', 'F', 'I', 'B', 'C4', 'ARG', 'SIM']
ignore = ['W291', 'W292', 'W293']

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.litellm]
provider = "groq"
api_key = "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"




